+++ pymd/backend/app/models/__init__.py
@@ -1,5 +1,5 @@
 """SQLAlchemy Models"""
-from pymd.backend.app.models.user import User
+from pymd.backend.app.models.user import User, UserRole
 from pymd.backend.app.models.document import Document
 from pymd.backend.app.models.session import Session
 from pymd.backend.app.models.user_settings import UserSettings
@@ -9,6 +9,7 @@ from pymd.backend.app.models.environment_execution import EnvironmentExecution, 
 
 __all__ = [
     "User",
+    "UserRole",
     "Document",
     "Session",
     "UserSettings",